refactor; cleanup tests
