+++ cognite_toolkit/_cdf_tk/loaders/_resource_loaders/transformation_loaders.py
@@ -89,7 +89,7 @@ class TransformationLoader(
     folder_name = "transformations"
     filename_pattern = (
         # Matches all yaml files except file names whose stem contain *.schedule. or .Notification
-        r"^(?!.*schedule.*|.*notification).*$"
+        r"^(?!.*schedule.*|.*\.notification$).*$"
     )
     resource_cls = Transformation
     resource_write_cls = TransformationWrite
